Replace decimal ID bytes (blocks/maps) with hex values
Will make it easier to compare directly with the tape bytes

+++ tape/block.go
@@ -2,7 +2,7 @@ package tape
 
 // Block is an interface for TZX data blocks
 type Block interface {
-	Id() int
+	Id() uint8
 	Name() string
 	ToString() string
 }
